Add ENTITY-MIB description -> ID mapping
The ENTITY-MIB is used in lots of equipment, but it can be quite
cumbersome to map all items manually, and the IDs might not make much
sense when printed to the user.

This adds the --entityname or -E option to enter an item description,
which then is used to extract ENTITY-MIB::entPhysicalDescr items and
use the matching item ID, adding it to the OID(s) specified (with the
assumption that all OID(s) specified are ENTITY items, such as
ENTITY-STATE-MIB::entStateOper).

To minimize impact, the result of the ENTITY-MIB::entPhysicalDescr
lookup is cached in a host-specific file. The directory for this is
currently hard-coded to the common Icinga/Nagios /run directories,
honoring the TMPDIR environment variable as an override.

Additionally, some other minor fixes are included.