Add option --minresults / -r for specifying expected values when walking
Some devices have the bad habit of just not listing devices that has
stopped working, instead of reporting an error status. An example of
this the IBM TS4500 tape library that tends to just ignore a tape drive
when communication has stopped working entirely.

Adding an option to specifying the minimum number of results/values
expected allows at least getting a warning when this happens.